+++ auth_backend/utils/auth_params.py
@@ -13,13 +13,3 @@ def get_auth_params(user_id: int, auth_method: str, session: Session) -> dict[st
     for method in methods:
         retval[method.param] = method
     return retval
-
-
-def get_users_auth_params(auth_method: str, session: Session) -> dict[int, dict[str, AuthMethod]]:
-    retval = {}
-    methods: list[AuthMethod] = AuthMethod.query(session=session).filter(AuthMethod.auth_method == auth_method).all()
-    for method in methods:
-        if method.user_id not in retval:
-            retval[method.user_id] = {}
-        retval[method.user_id][method.param] = method
-    return retval
